Tweak UDO implementation
UDOs are now supported by the lexer,. This simplifies parsing logic and
eliminates the potential for weird parse behaviour and a whole host of
potential problems that are difficult to find, fix, and maintain, as
well as removing the need for difficult to maintain kludges to handle
built-in operators and UDOs. Although this does introduce a couple new
minor side effects, they are less severe than the problems this tweak
fixes.

Punctuation/operators have been further segregated in the lexer.