Scrap old implementation of destructuring.
The new implementation of destructuring is handled by two new
expressions to avoid transforming the AST, so destructuring is handled
at compile-time rather than parse-time.

This commit also brings with it a change to load undefined when
destructuring more values than an array contains, instead of throwing an
index out of bounds exception.